+++ src/main/java/com/umc/product/organization/adapter/in/web/ChapterQueryController.java
@@ -1,9 +1,9 @@
 package com.umc.product.organization.adapter.in.web;
 
 import com.umc.product.global.security.annotation.Public;
-import com.umc.product.organization.adapter.in.web.dto.response.ChapterListResponse;
-import com.umc.product.organization.adapter.in.web.dto.response.ChapterListResponse.ChapterItem;
-import com.umc.product.organization.adapter.in.web.dto.response.ChapterWithSchoolsResponse;
+import com.umc.product.organization.adapter.in.web.dto.response.chapter.ChapterListResponse;
+import com.umc.product.organization.adapter.in.web.dto.response.chapter.ChapterListResponse.ChapterItem;
+import com.umc.product.organization.adapter.in.web.dto.response.chapter.ChapterWithSchoolsResponse;
 import com.umc.product.organization.adapter.in.web.swagger.ChapterQueryControllerApi;
 import com.umc.product.organization.application.port.in.query.GetChapterUseCase;
 import com.umc.product.organization.application.port.in.query.dto.chapter.ChapterInfo;
