+++ src/main/java/com/umc/product/organization/adapter/in/web/GisuQueryController.java
@@ -2,10 +2,10 @@ package com.umc.product.organization.adapter.in.web;
 
 import com.umc.product.global.response.PageResponse;
 import com.umc.product.global.security.annotation.Public;
-import com.umc.product.organization.adapter.in.web.dto.response.ActiveGisuResponse;
-import com.umc.product.organization.adapter.in.web.dto.response.GisuNameListResponse;
-import com.umc.product.organization.adapter.in.web.dto.response.GisuPageResponse;
-import com.umc.product.organization.adapter.in.web.dto.response.GisuResponse;
+import com.umc.product.organization.adapter.in.web.dto.response.gisu.ActiveGisuResponse;
+import com.umc.product.organization.adapter.in.web.dto.response.gisu.GisuNameListResponse;
+import com.umc.product.organization.adapter.in.web.dto.response.gisu.GisuPageResponse;
+import com.umc.product.organization.adapter.in.web.dto.response.gisu.GisuResponse;
 import com.umc.product.organization.adapter.in.web.swagger.AdminGisuQueryControllerApi;
 import com.umc.product.organization.application.port.in.query.GetGisuUseCase;
 import lombok.RequiredArgsConstructor;
