+++ src/main/java/com/umc/product/organization/adapter/in/web/StudyGroupQueryController.java
@@ -6,9 +6,9 @@ import com.umc.product.authorization.domain.ResourceType;
 import com.umc.product.global.response.CursorResponse;
 import com.umc.product.global.security.MemberPrincipal;
 import com.umc.product.global.security.annotation.CurrentMember;
-import com.umc.product.organization.adapter.in.web.dto.response.StudyGroupMemberResponse;
-import com.umc.product.organization.adapter.in.web.dto.response.StudyGroupNameResponse;
-import com.umc.product.organization.adapter.in.web.dto.response.StudyGroupResponse;
+import com.umc.product.organization.adapter.in.web.dto.response.studygroup.StudyGroupMemberResponse;
+import com.umc.product.organization.adapter.in.web.dto.response.studygroup.StudyGroupNameResponse;
+import com.umc.product.organization.adapter.in.web.dto.response.studygroup.StudyGroupResponse;
 import com.umc.product.organization.adapter.in.web.swagger.StudyGroupQueryControllerApi;
 import com.umc.product.organization.application.port.in.query.GetStudyGroupUseCase;
 import com.umc.product.organization.application.port.in.query.dto.studygroup.StudyGroupInfo;
