+++ src/main/java/com/umc/product/organization/adapter/in/web/dto/response/school/SchoolDetailResponse.java
@@ -1,4 +1,4 @@
-package com.umc.product.organization.adapter.in.web.dto.response;
+package com.umc.product.organization.adapter.in.web.dto.response.school;
 
 import com.umc.product.organization.application.port.in.query.dto.school.SchoolDetailInfo;
 import com.umc.product.organization.domain.enums.SchoolLinkType;
