+++ src/main/java/com/umc/product/organization/adapter/in/web/swagger/SchoolQueryControllerApi.java
@@ -1,9 +1,9 @@
 package com.umc.product.organization.adapter.in.web.swagger;
 
-import com.umc.product.organization.adapter.in.web.dto.response.SchoolDetailResponse;
-import com.umc.product.organization.adapter.in.web.dto.response.SchoolLinkResponse;
-import com.umc.product.organization.adapter.in.web.dto.response.SchoolNameListResponse;
-import com.umc.product.organization.adapter.in.web.dto.response.UnassignedSchoolListResponse;
+import com.umc.product.organization.adapter.in.web.dto.response.school.SchoolDetailResponse;
+import com.umc.product.organization.adapter.in.web.dto.response.school.SchoolLinkResponse;
+import com.umc.product.organization.adapter.in.web.dto.response.school.SchoolNameListResponse;
+import com.umc.product.organization.adapter.in.web.dto.response.school.UnassignedSchoolListResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
