+++ src/main/java/com/umc/product/organization/adapter/in/web/swagger/StudyGroupQueryControllerApi.java
@@ -2,9 +2,9 @@ package com.umc.product.organization.adapter.in.web.swagger;
 
 import com.umc.product.global.response.CursorResponse;
 import com.umc.product.global.security.MemberPrincipal;
-import com.umc.product.organization.adapter.in.web.dto.response.StudyGroupMemberResponse;
-import com.umc.product.organization.adapter.in.web.dto.response.StudyGroupNameResponse;
-import com.umc.product.organization.adapter.in.web.dto.response.StudyGroupResponse;
+import com.umc.product.organization.adapter.in.web.dto.response.studygroup.StudyGroupMemberResponse;
+import com.umc.product.organization.adapter.in.web.dto.response.studygroup.StudyGroupNameResponse;
+import com.umc.product.organization.adapter.in.web.dto.response.studygroup.StudyGroupResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
